Closes #2
