+++ app/src/main/java/com/example/ryanzhouold/bakingandroid/RecipeListContract.java
@@ -1,10 +1,12 @@
 package com.example.ryanzhouold.bakingandroid;
 
+import android.support.v4.app.Fragment;
+
 public interface RecipeListContract {
     interface View{
-        void showRecipes();
+        void showRecipes(String data);
     }
     interface Presenter{
-        void onClickRecipe();
+        void loadRecipes();
     }
 }
